+++ internal/cmd/marketplace/marketplace_item.go
@@ -13,29 +13,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package cmd
+package marketplace
 
-import (
-	"github.com/mia-platform/miactl/internal/clioptions"
-	"github.com/mia-platform/miactl/internal/cmd/marketplace"
-	"github.com/spf13/cobra"
-)
+import "github.com/mia-platform/miactl/internal/encoding"
 
-func MarketplaceCmd(options *clioptions.CLIOptions) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "marketplace",
-		Short: "View and manage marketplace items",
+func Unmarshal(data []byte, encodingFormat string) (Item, error) {
+	var r Item
+
+	if err := encoding.UnmarshalData(data, encodingFormat, &r); err != nil {
+		return Item{}, err
 	}
-
-	// add cmd flags
-	flags := cmd.PersistentFlags()
-	options.AddConnectionFlags(flags)
-	options.AddCompanyFlags(flags)
-	options.AddContextFlags(flags)
-
-	// add sub commands
-	cmd.AddCommand(marketplace.ListCmd(options))
-	cmd.AddCommand(marketplace.GetCmd(options))
-
-	return cmd
+	return r, nil
 }
+
+func (r *Item) Marshal(encodingFormat string) ([]byte, error) {
+	return encoding.MarshalData(r, encodingFormat, encoding.MarshalOptions{Indent: true})
+}
+
+// we use a map[string]interface{} to represent the item
+// this will allow us to avoid to change the code in case of a change in the resource structure
+type Item map[string]interface{}